data/requirements.txt
requests==2.32.3
mysql-connector-python
